Fixed poisson intercept to work with MPI. Merged MPI tests into one, and added enet/poisson tests. Fixed warnings in uoi lasso tests.
