+++ bin/test_mpi.sh
@@ -1,2 +1,2 @@
-mpiexec -n 4 pytest -sv tests/test_mpi
-mpiexec -n 49 pytest -sv tests/test_mpi
+mpiexec -n 4 pytest -sv $1
+mpiexec -n 49 pytest -sv $1
